refactor(preprocessor): reorganize core directory structure
- Move filter, process, profiles, variant to core/preprocessing/
- Move llm.ts, summarize/, query/ to core/ai/command/
- Update all import paths

Directory structure:
  src/core/
    preprocessing/   # Data transformation
      filter.ts
      process.ts
      profiles.ts
      variant.ts
    ai/              # AI operations
      llm.ts
      command/
        summarize/
        query/
